+++ apps/data-pipeline/tests/slow/test_stage4b_parametric_id.py
@@ -35,7 +35,7 @@ def _support_runtime() -> ObservationSupportRuntime:
         interval_prev_coeffs=np.array([[[0.0], [0.0], [0.0]], [[0.0], [0.0], [0.5]]]),
         interval_curr_coeffs=np.array([[[0.0], [0.0], [0.0]], [[0.0], [0.0], [0.5]]]),
         interval_weights=np.array([[[0.0], [0.0], [0.0]], [[0.0], [0.0], [1.0]]]),
-        emission_slot_indices=np.array([[-1, -1, -1], [-1, -1, 0]], dtype=np.int32),
+        emission_slot_indices=np.array([[-1, -1, -1], [-1, -1, 0]], dtype=np.int64),
     )
